refactor(mobile): CertificateBorder 공용 컴포넌트 분리
AchievementSummaryCard에서 CertificateBorder를 별도 파일로 추출하여
상세 페이지와 리스트 요약 카드에서 공유
